Add files via upload
The python code is to text mine all the phytochemical data from the IMPPAT database ( Indian Medicinal Plants, Phytochemistry And Therapeutics) at once and download all the phytochemicals in SDF format from the PubChem database in a folder. 
The code will be useful in the bulk download of small compounds from various databases for high-throughput virtual screening.